requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
Django==2.2.10
django-bootstrap4==2.3.1
django-cleanup==5.1.0
django-cors-headers==3.6.0
django-ranged-response==0.2.0
django-simple-captcha==0.5.13
djangorestframework==3.12.2
easy-thumbnails==2.7.1
gunicorn==20.0.4
importlib-metadata==2.1.1
Pillow==8.1.0
pytz==2020.5
six==1.15.0
soupsieve==2.1
sqlparse==0.4.1
zipp==3.4.0
